refactor(extension): create a feed component
The feed component is a standalone component which presents a complete feed of posts.
The component takes care of bookmarking, reporting and all the interaction needed with the posts.
The feed takes into account the user preferences (cards/insane).
