+++ packages/extension/__tests__/Home.js
@@ -3,24 +3,20 @@ import Vuex from 'vuex';
 import svgicon from 'vue-svgicon';
 import VTooltip from 'v-tooltip';
 import DaHome from '../src/routes/Home.vue';
-import DaCardPost from '../../components/src/components/DaCardPost.vue';
-import DaInsanePost from '../../components/src/components/DaInsanePost.vue';
 import DaContext from '../../components/src/components/DaContext.vue';
-import DaCardPlaceholder from '../../components/src/components/DaCardPlaceholder.vue';
 import DaHeader from '../src/components/DaHeader.vue';
 import DaSidebar from '../src/components/DaSidebar.vue';
+import DaFeed from '../src/components/DaFeed.vue';
 
 const localVue = createLocalVue();
 
 localVue.use(Vuex);
 localVue.use(svgicon);
 localVue.use(VTooltip);
-localVue.component('da-card-post', DaCardPost);
-localVue.component('da-insane-post', DaInsanePost);
 localVue.component('da-context', DaContext);
-localVue.component('da-card-placeholder', DaCardPlaceholder);
 localVue.component('da-header', DaHeader);
 localVue.component('da-sidebar', DaSidebar);
+localVue.component('da-feed', DaFeed);
 
 let feed;
 let ui;
@@ -34,57 +30,24 @@ beforeEach(() => {
   feed = {
     namespaced: true,
     state: {
-      publications: [{
-        'id': 'devto',
-        'name': 'DEV',
-        'image': 'https://res.cloudinary.com/daily-now/image/upload/t_logo,f_auto/v1/logos/devto',
-        'enabled': true,
-        'twitter': ''
-      }],
-      posts: [{
-        'id': '0b14843f08279a41ea69c1edd35a2c16',
-        'title': 'Four Firefox Dev-Tools I use daily',
-        'url': 'https://app.dailynow.co/r/0b14843f08279a41ea69c1edd35a2c16',
-        'publishedAt': '2019-06-12T00:51:56.000Z',
-        'createdAt': '2019-06-12T18:54:48.000Z',
-        'publication': {
-          'id': 'devto',
-          'name': 'DEV',
-          'image': 'https://res.cloudinary.com/daily-now/image/upload/t_logo,f_auto/v1/logos/devto'
-        },
-        'views': 130,
-        'tags': [
-          'webdev',
-          'beginners',
-          'css'
-        ],
-        'readTime': 4,
-        'image': 'https://res.cloudinary.com/daily-now/image/upload/f_auto,q_auto/v1/posts/5fdc29751a3ca792d72e6456016faf0f',
-        'placeholder': 'data:image/jpeg;base64,/9j/2wBDAAYEBQYFBAYGBQYHBwYIChAKCgkJChQODwwQFxQYGBcUFhYaHSUfGhsjHBYWICwgIyYnKSopGR8tMC0oMCUoKSj/2wBDAQcHBwoIChMKChMoGhYaKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCgoKCj/wAARCAAFAAoDASIAAhEBAxEB/8QAFgABAQEAAAAAAAAAAAAAAAAAAAMF/8QAGhAAAgIDAAAAAAAAAAAAAAAAAAECERMxUf/EABUBAQEAAAAAAAAAAAAAAAAAAAYH/8QAFREBAQAAAAAAAAAAAAAAAAAAAAH/2gAMAwEAAhEDEQA/ANWcpbTVcojkYATUyR//2Q==',
-        'size': 'small'
-      }]
+      publications: [],
+      posts: [],
     },
     mutations: {
       setPublications: jest.fn()
     },
-    actions: {
-      setFilter: jest.fn()
-    },
     getters: {
       feed: state => state['posts'],
-      showAd: jest.fn()
+      showAd: jest.fn(),
     }
   };
 
   ui = {
     namespaced: true,
-    enableCardAnimations: true,
-    instructionsStep: 0,
     state: {
       insaneMode: false,
     },
     mutations: {
-      setInsaneMode: jest.fn(),
       setDndModeTime: jest.fn(),
       setShowDndMenu: jest.fn(),
     },
@@ -116,37 +79,6 @@ beforeEach(() => {
 
   window.requestIdleCallback = jest.fn();
   window.IntersectionObserver = jest.fn();
-});
-
-it('should dispatch "setFilter" with publication filter', (done) => {
-  const wrapper = mount(DaHome, { store, localVue });
-  const expectedPublication = feed.state.posts[0].publication;
-  wrapper.vm.$nextTick(() => {
-    wrapper
-    .find('.card__footer__publication').trigger('click');
-    expect(feed.actions.setFilter)
-    .toBeCalledWith(expect.anything(), {
-      type: 'publication',
-      info: expectedPublication,
-    }, undefined);
-    done();
-  });
-});
-
-it('should dispatch "setFilter" with publication filter when in insane mode', (done) => {
-  const wrapper = mount(DaHome, { store, localVue });
-  const expectedPublication = feed.state.posts[0].publication;
-  store.state.ui.insaneMode = true;
-  wrapper.vm.$nextTick(() => {
-    wrapper
-    .find('.insane__publication').trigger('click');
-    expect(feed.actions.setFilter)
-    .toBeCalledWith(expect.anything(), {
-      type: 'publication',
-      info: expectedPublication,
-    }, undefined);
-    done();
-  });
 });
 
 it('should commit "setDndModeTime" when "For 1 Hour" or "Until Tomorrow" is clicked', () => {
